refactor(bookmark): update sorting logic to use Map for parameters and change sort type from Long to Int across relevant entities and services

+++ bookmarkify-api/src/main/kotlin/top/tcyeee/bookmarkify/entity/Response.kt
@@ -132,7 +132,7 @@ data class UserPreferenceVO(
 data class UserLayoutNodeVO(
     @field:Schema(description = "节点ID") val id: String = IdUtil.fastUUID(),
     @field:Schema(description = "父节点ID") val parentId: String? = null,
-    @field:Schema(description = "排序") val sort: Long = Long.MAX_VALUE,
+    @field:Schema(description = "排序") val sort: Int = Int.MAX_VALUE,
     @field:Schema(description = "节点类型") val type: NodeTypeEnum = NodeTypeEnum.BOOKMARK,
     @field:Schema(description = "节点(文件夹)名称") val name: String? = null,
